Dev: Added some cp0 registers

+++ src/mm/tlb.rs
@@ -45,7 +45,7 @@ pub fn passive_alloc(va: VA, pgdir: PageDirectory, asid: usize) {
 }
 
 /// This function returns (entrylo0, entrylo1).
-pub fn do_tlb_refill(va: VA, asid: usize) -> (usize, usize) {
+pub fn do_tlb_refill(va: VA, asid: usize) -> (u32, u32) {
     tlb_invalidate(asid, va);
     // TODO:
     (0, 0)
